em, a lot has been changed or added

+++ tst/test_pinocchio.c
@@ -2,9 +2,10 @@
 #include "pinocchio.h"
 #include "test_pinocchio.h"
 
-void test_cindex(char *s, char c, signed long long int n, size_t start, signed long long int must, char *file, size_t line)
+void test_cindex(char *s, char c, signed long long int n, size_t start, size_t must, char *file, size_t line)
 {
-   signed long long int just = cindex(s,c,n,start);
+   /* size_t just = int_cindex(s,c,n,start); */
+   size_t just = cindex(s,c,n,start);
    if(isEqualSInt(just, must))
    {
       PASSED;
@@ -12,8 +13,8 @@ void test_cindex(char *s, char c, signed long long int n, size_t start, signed l
    else
    {
       FAILED;
-      fprintf(stderr, "\tmust: %lld\n", must);
-      fprintf(stderr, "\tjust: %lld\n", just);
+      fprintf(stderr, "\tmust: %zu\n", must);
+      fprintf(stderr, "\tjust: %zu\n", just);
    }
 }
