+++ tst/test_pinocchio.h
@@ -2,4 +2,4 @@
 #include "pinocchio.h"
 
 
-void test_cindex(char *s, char c, signed long long int n, size_t start, signed long long int must, char *file, size_t line);
+void test_cindex(char *s, char c, signed long long int n, size_t start, size_t must, char *file, size_t line);
